+++ src/js/utils.js
@@ -6,6 +6,8 @@ if ( typeof exports === 'object' && typeof exports.nodeName !== 'string' && type
 define( function( require, exports, module ) {
     'use strict';
 
+    var cookies;
+
     /**
      * Parses an Expression to extract all function calls and theirs argument arrays.
      *
@@ -98,11 +100,39 @@ define( function( require, exports, module ) {
         return !isNaN( parseFloat( n ) ) && isFinite( n );
     }
 
+    function readCookie( name ) {
+        var c;
+        var C;
+        var i;
+
+        if ( cookies ) {
+            return cookies[ name ];
+        }
+
+        c = document.cookie.split( '; ' );
+        cookies = {};
+
+        for ( i = c.length - 1; i >= 0; i-- ) {
+            C = c[ i ].split( '=' );
+            // decode URI
+            C[ 1 ] = decodeURIComponent( C[ 1 ] );
+            // if cookie is signed (using expressjs/cookie-parser/), extract value
+            if ( C[ 1 ].substr( 0, 2 ) === 's:' ) {
+                C[ 1 ] = C[ 1 ].slice( 2 );
+                C[ 1 ] = C[ 1 ].slice( 0, C[ 1 ].lastIndexOf( '.' ) );
+            }
+            cookies[ C[ 0 ] ] = decodeURIComponent( C[ 1 ] );
+        }
+
+        return cookies[ name ];
+    }
+
     module.exports = {
         parseFunctionFromExpression: parseFunctionFromExpression,
         stripQuotes: stripQuotes,
         getFilename: getFilename,
         toArray: toArray,
-        isNumber: isNumber
+        isNumber: isNumber,
+        readCookie: readCookie
     };
 } );