+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Ability to pass session properties (metadata) when instantiating form.
+
+
 [4.24.2] - 2017-03-03
 ---------------------
 ##### Fixed